:truck: rename: _layout folder name into layout
